quick-compare/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
playwright==1.47.0
pydantic==2.9.2
aiosqlite==0.20.0
python-dotenv==1.0.1
pycryptodome==3.20.0
